backend/requirements.txt
django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-decouple==3.8
